iptv/requirements.txt
selenium==4.19.0
selenium-wire==5.1.0
webdriver-manager==4.0.2
proxy.py==2.4.3
GitPython==3.1.43
python-dotenv==1.0.1